+++ tools/genSettings.py
@@ -66,7 +66,7 @@ namespace SettingsHelper
 
 struct SettingInfo
 {
-    QString name, desc, type, ui;
+    QString name, desc, type, ui, tip, help;
     QStringList old;
     QVariant def;
     QVariant param;
@@ -91,7 +91,15 @@ const SettingInfo settingInfo[] =
             ui = t["ui"]
         else:
             ui = ""
-        setting_info.write(f"    {{{json.dumps(name)}, {json.dumps(desc)}, \"{typename}\", \"{ui}\", {{")
+        if "tip" in t:
+            tip = t["tip"]
+        else:
+            tip = ""
+        if "help" in t:
+            hlp = t["hlp"]
+        else:
+            hlp = ""
+        setting_info.write(f"    {{{json.dumps(name)}, {json.dumps(desc)}, \"{typename}\", \"{ui}\", {json.dumps(tip)}, {json.dumps(hlp)}, {{")
         if "old" in t:
             olds = ""
             for s in t["old"]: