+++ doc/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - Add Ctrl+Shift+D for duplicating the current line or the selection. (#237)
+- Add tooltips for many widgets. (#211)
 
 ### Changed